stack create command with copy optional flag
